+++ history.md
@@ -1,3 +1,8 @@
+1.0.3 / 2022-01-15
+===================
+
+    * Fix: Issue where if params had extra url they would become corrupt
+
 1.0.2 / 2022-01-15
 ===================